Refactor: #175 모델 개인화 모듈 네이밍 변경 및 리팩토링
What?
- Personalizater라는 네이밍이 어색해서 Personalizer로 변경.
- 불필요한 Print문, Print를 위한 메서드 제거.
- 모델 저장 책임 ModelRepository에 위임.
- 프로토콜 네이밍 포함.
